+++ packages/integrations/mdx/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @astrojs/mdx
 
+## 5.0.0-beta.9
+
+### Patch Changes
+
+- Updated dependencies [[`1fa4177`](https://github.com/withastro/astro/commit/1fa41779c458123f707940a5253dbe6e540dbf7d)]:
+  - @astrojs/markdown-remark@7.0.0-beta.8
+
 ## 5.0.0-beta.8
 
 ### Major Changes